+++ MissionMonitor/MissionMonitor.cs
@@ -944,113 +944,6 @@ namespace EddiMissionMonitor
             }
         }
 
-        public void CancelRoute()
-        {
-            missionsRouteList = null;
-            missionsRouteDistance = 0;
-            writeMissions();
-
-            // Clear destination variables
-            UpdateDestinationData(null, 0);
-
-            EDDI.Instance.enqueueEvent(new RouteDetailsEvent(DateTime.Now, "cancel", null, missionsRouteList, 0, 0, missionsRouteDistance, null));
-        }
-
-        public string GetExpiringRoute()
-        {
-            string expiringSystem = null;
-            decimal expiringDistance = 0;
-            long expiringSeconds = 0;
-            List<long> missionids = new List<long>();       // List of mission IDs for the next system
-
-            if (missionsCount > 0)
-            {
-                StarSystem curr = EDDI.Instance?.CurrentStarSystem;
-                StarSystem dest = new StarSystem();             // Destination star system
-
-                foreach (Mission mission in missions.Where(m => m.statusEDName == "Active").ToList())
-                {
-                    if (expiringSeconds == 0 || mission.expiryseconds < expiringSeconds)
-                    {
-                        expiringSeconds = mission.expiryseconds ?? 0;
-                        expiringSystem = mission.destinationsystem;
-                    }
-                }
-
-                dest = StarSystemSqLiteRepository.Instance.GetOrCreateStarSystem(expiringSystem, true);
-                expiringDistance = CalculateDistance(curr, dest);
-                missionsRouteList = expiringSystem;
-                missionsRouteDistance = expiringDistance;
-                writeMissions();
-
-                // Get mission IDs for 'expiring' system
-                missionids = GetSystemMissionIds(expiringSystem);
-
-                // Set destination variables
-                UpdateDestinationData(expiringSystem, expiringDistance);
-
-            }
-            EDDI.Instance.enqueueEvent(new RouteDetailsEvent(DateTime.Now, "expiring", expiringSystem, missionsRouteList, expiringSeconds, expiringDistance, missionsRouteDistance, missionids));
-            return expiringSystem;
-        }
-
-        public string GetFarthestRoute()
-        {
-            // Missions Route Event variables
-            string farthestSystem = null;
-            decimal farthestDistance = 0;
-            List<long> missionids = new List<long>();       // List of mission IDs for the next system
-
-            if (missionsCount > 0)
-            {
-                StarSystem curr = EDDI.Instance?.CurrentStarSystem;
-                StarSystem dest = new StarSystem();             // Destination star system
-
-                SortedList<decimal, string> farthestList = new SortedList<decimal, string>();
-                foreach (Mission mission in missions.Where(m => m.statusEDName == "Active").ToList())
-                {
-                    if (mission.destinationsystems != null && mission.destinationsystems.Any())
-                    {
-                        foreach (DestinationSystem system in mission.destinationsystems)
-                        {
-                            dest = StarSystemSqLiteRepository.Instance.GetOrCreateStarSystem(system.name, true);
-                            decimal distance = CalculateDistance(curr, dest);
-                            if (!farthestList.ContainsKey(distance))
-                            {
-                                farthestList.Add(distance, system.name);
-
-                            }
-                        }
-                    }
-                    else if (mission.destinationsystem != string.Empty)
-                    {
-                        dest = StarSystemSqLiteRepository.Instance.GetOrCreateStarSystem(mission.destinationsystem, true);
-                        decimal distance = CalculateDistance(curr, dest);
-                        if (!farthestList.ContainsKey(distance))
-                        {
-                            farthestList.Add(distance, mission.destinationsystem);
-                        }
-                    }
-                }
-
-                // Farthest system is last in the list
-                farthestSystem = farthestList.Values.LastOrDefault();
-                farthestDistance = farthestList.Keys.LastOrDefault();
-                missionsRouteList = farthestSystem;
-                missionsRouteDistance = farthestDistance;
-                writeMissions();
-
-                // Get mission IDs for 'farthest' system
-                missionids = GetSystemMissionIds(farthestSystem);
-
-                // Set destination variables
-                UpdateDestinationData(farthestSystem, farthestDistance);
-
-            }
-            EDDI.Instance.enqueueEvent(new RouteDetailsEvent(DateTime.Now, "farthest", farthestSystem, missionsRouteList, missionids.Count(), farthestDistance, missionsRouteDistance, missionids));
-            return farthestSystem;
-        }
-
         public string GetMostRoute(string homeSystem = null)
         {
             // Missions Route Event variables
@@ -1147,63 +1040,6 @@ namespace EddiMissionMonitor
             }
             EDDI.Instance.enqueueEvent(new RouteDetailsEvent(DateTime.Now, "most", mostSystem, missionsRouteList, mostCount, mostDistance, missionsRouteDistance, missionids));
             return mostSystem;
-        }
-
-        public string GetNearestRoute()
-        {
-            // Missions Route Event variables
-            string nearestSystem = null;
-            decimal nearestDistance = 0;
-            List<long> missionids = new List<long>();       // List of mission IDs for the next system
-
-            if (missionsCount > 0)
-            {
-                StarSystem curr = EDDI.Instance?.CurrentStarSystem;     // Current star system
-                StarSystem dest = new StarSystem();                     // Destination star system
-
-                SortedList<decimal, string> nearestList = new SortedList<decimal, string>();
-                foreach (Mission mission in missions.Where(m => m.statusEDName == "Active").ToList())
-                {
-                    if (mission.destinationsystems != null && mission.destinationsystems.Any())
-                    {
-                        foreach (DestinationSystem system in mission.destinationsystems)
-                        {
-                            dest = StarSystemSqLiteRepository.Instance.GetOrCreateStarSystem(system.name, true);
-                            decimal distance = CalculateDistance(curr, dest);
-                            if (!nearestList.ContainsKey(distance))
-                            {
-                                nearestList.Add(distance, system.name);
-
-                            }
-                        }
-                    }
-                    else if (mission.destinationsystem != string.Empty)
-                    {
-                        dest = StarSystemSqLiteRepository.Instance.GetOrCreateStarSystem(mission.destinationsystem, true);
-                        decimal distance = CalculateDistance(curr, dest);
-                        if (!nearestList.ContainsKey(distance))
-                        {
-                            nearestList.Add(distance, mission.destinationsystem);
-                        }
-                    }
-                }
-
-                // Nearest system is first in the list
-                nearestSystem = nearestList.Values.FirstOrDefault();
-                nearestDistance = nearestList.Keys.FirstOrDefault();
-                missionsRouteList = nearestSystem;
-                missionsRouteDistance = nearestDistance;
-                writeMissions();
-
-                // Get mission IDs for 'nearest' system
-                missionids = GetSystemMissionIds(nearestSystem);
-
-                // Set destination variables
-                UpdateDestinationData(nearestSystem, nearestDistance);
-
-            }
-            EDDI.Instance.enqueueEvent(new RouteDetailsEvent(DateTime.Now, "nearest", nearestSystem, missionsRouteList, missionids.Count(), nearestDistance, missionsRouteDistance, missionids));
-            return nearestSystem;
         }
 
         public string GetMissionsRoute(string homeSystem = null)
@@ -1413,66 +1249,7 @@ namespace EddiMissionMonitor
             return missionsRouteList?.Split('_')[0];
         }
 
-        public string SetNextRoute()
-        {
-            string nextSystem = missionsRouteList?.Split('_')[0];
-            decimal nextDistance = 0;
-            int count = 0;
-            List<long> missionids = new List<long>();       // List of mission IDs for the next system
-
-            if (nextSystem != null)
-            {
-                StarSystem curr = EDDI.Instance?.CurrentStarSystem;
-                StarSystem dest = StarSystemSqLiteRepository.Instance.GetOrFetchStarSystem(nextSystem, true);
-
-                if (dest != null && nextSystem != curr.name)
-                {
-                    nextDistance = CalculateDistance(curr, dest);
-                }
-                count = missionsRouteList.Split('_').Count();
-
-                // Get mission IDs for 'next' system
-                missionids = GetSystemMissionIds(nextSystem);
-
-                // Set destination variables
-                UpdateDestinationData(nextSystem, nextDistance);
-            }
-            EDDI.Instance.enqueueEvent(new RouteDetailsEvent(DateTime.Now, "next", nextSystem, missionsRouteList, count, nextDistance, missionsRouteDistance, missionids));
-            return nextSystem;
-        }
-
-        public string SetRoute(string system)
-        {
-            string destination = null;
-            decimal distance = 0;
-            List<long> missionids = new List<long>();       // List of mission IDs for the next system
-
-            if (system != null)
-            {
-                StarSystem curr = EDDI.Instance?.CurrentStarSystem;
-                StarSystem dest = StarSystemSqLiteRepository.Instance.GetOrCreateStarSystem(system, true);
-
-                if (dest != null && system != curr.name)
-                {
-                    distance = CalculateDistance(curr, dest);
-                    destination = dest.name;
-                }
-                missionsRouteList = destination;
-                missionsRouteDistance = distance;
-                writeMissions();
-
-                // Get mission IDs for 'next' system
-                missionids = GetSystemMissionIds(destination);
-
-                // Set destination variables
-                UpdateDestinationData(destination, distance);
-
-            }
-            EDDI.Instance.enqueueEvent(new RouteDetailsEvent(DateTime.Now, "set", destination, missionsRouteList, 1, distance, missionsRouteDistance, missionids));
-            return destination;
-        }
-
-        public string UpdateMissionsRoute(string updateSystem = null)
+       public string UpdateMissionsRoute(string updateSystem = null)
         {
             bool update;
             string nextSystem = null;
