+++ SpeechResponder/ScriptResolver.cs
@@ -6,9 +6,11 @@ using Cottle.Stores;
 using Cottle.Values;
 using Eddi;
 using EddiCargoMonitor;
+using EddiCrimeMonitor;
 using EddiDataDefinitions;
 using EddiDataProviderService;
 using EddiMissionMonitor;
+using EddiNavigationService;
 using EddiShipMonitor;
 using EddiSpeechService;
 using GalnetMonitor;
@@ -656,17 +658,22 @@ namespace EddiSpeechResponder
                 {
                     case "cancel":
                         {
-                            ((MissionMonitor)EDDI.Instance.ObtainMonitor("Mission monitor"))?.CancelRoute();
+                            Navigation.Instance.CancelRoute();
                         }
                         break;
                     case "expiring":
                         {
-                            result = ((MissionMonitor)EDDI.Instance.ObtainMonitor("Mission monitor"))?.GetExpiringRoute();
+                            result = Navigation.Instance.GetExpiringRoute();
+                        }
+                        break;
+                    case "facillitator":
+                        {
+                            result = Navigation.Instance.GetFacillitatorRoute();
                         }
                         break;
                     case "farthest":
                         {
-                            result = ((MissionMonitor)EDDI.Instance.ObtainMonitor("Mission monitor"))?.GetFarthestRoute();
+                            result = Navigation.Instance.GetFarthestRoute();
                         }
                         break;
                     case "most":
@@ -683,12 +690,12 @@ namespace EddiSpeechResponder
                         break;
                     case "nearest":
                         {
-                            result = ((MissionMonitor)EDDI.Instance.ObtainMonitor("Mission monitor"))?.GetNearestRoute();
+                            result = Navigation.Instance.GetNearestRoute();
                         }
                         break;
                     case "next":
                         {
-                            result = ((MissionMonitor)EDDI.Instance.ObtainMonitor("Mission monitor"))?.SetNextRoute();
+                            result = Navigation.Instance.SetNextRoute();
                         }
                         break;
                     case "route":
@@ -705,18 +712,18 @@ namespace EddiSpeechResponder
                         break;
                     case "set":
                         {
-                            result = ((MissionMonitor)EDDI.Instance.ObtainMonitor("Mission monitor"))?.SetRoute(values[1].AsString);
+                            result = Navigation.Instance.SetRoute(values[1].AsString);
                         }
                         break;
                     case "source":
                         {
                             if (values.Count == 2)
                             {
-                                result = ((CargoMonitor)EDDI.Instance.ObtainMonitor("Cargo monitor"))?.GetSourceRoute(values[1].AsString);
+                                result = Navigation.Instance.GetSourceRoute(values[1].AsString);
                             }
                             else
                             {
-                                result = ((CargoMonitor)EDDI.Instance.ObtainMonitor("Cargo monitor"))?.GetSourceRoute();
+                                result = Navigation.Instance.GetSourceRoute();
                             }
                         }
                         break;
