+++ VoiceAttackResponder/EddiVoiceAttackResponder.csproj
@@ -87,6 +87,10 @@
       <Project>{c1d256ba-68b9-437a-8907-599c3a388c37}</Project>
       <Name>EddiCargoMonitor</Name>
     </ProjectReference>
+    <ProjectReference Include="..\EddiNavigationService\EddiNavigationService.csproj">
+      <Project>{86c23ebd-c979-4517-bb66-62e103ab13ef}</Project>
+      <Name>EddiNavigationService</Name>
+    </ProjectReference>
     <ProjectReference Include="..\EDDI\Eddi.csproj">
       <Project>{EC7BA042-A370-447F-8C3E-241358CEBCBB}</Project>
       <Name>Eddi</Name>
